requirements.txt
asgiref==3.4.1
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-disqus==0.5
Flask==2.0.1
Flask-Caching==1.10.1
html5lib==1.1
idna==2.10
isso==0.12.5
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
misaka==2.1.1
mock==4.0.3
oauthlib==3.1.1
packaging==21.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.2.0
